modules/location_producer/requirements.txt
grpcio-tools==1.46.0
kafka-python==2.0.2